backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
Pillow==11.3.0
pytesseract==0.3.13
requests==2.32.5
psycopg2-binary==2.9.10
